[refactor] ThreadItem 컴포넌트 호출 index 통하는것으로 변경
